+++ docs/content/docs/functions-jsx/bob.ts
@@ -1,0 +1,6 @@
+import type { ConfigFile } from "../../../../mod.ts";
+import functions from "./functions.tsx";
+
+export default <ConfigFile> {
+  functions,
+};